requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
SQLAlchemy>=2.0
asyncpg>=0.29
python-dotenv>=1.0
httpx>=0.27
pypdf>=4.0
